+++ src/ts/VEllipse.ts
@@ -283,6 +283,7 @@ export class VEllipse implements Intersectable, SVGSerializable {
    */
   tangentAt(angle: number, length?: number): Vector {
     const normal: Vector = this.normalAt(angle, length);
+    // TODO: cleanup
     // const normal: Vector = this.normalAt(angle - this.rotation, length);
     // Rotate the normal by 90 degrees, then it is the tangent.
     // normal.b.rotate(Math.PI / 2, normal.a);